All Project CRUDing functional
Added the visuals for index, update, and deleting a project.
All function work properly.

There was a bug in the update and delete at first because of a typo.
browser console indicated that it could not read the property of the
undefined ID. I checked the update & delete function and there was
indeed an issue with the assignment of the ID to be passed to the API.