Fix modal bug when it closes
When submitting a form in a modal, it would close but the class would
stay as .modal-open in the body which would prevent the ability to
scroll the page up and down. Fixed by making the overflow by inherit in
scss.

Made the nav bar float to the right rather than left which means I had
to reverse the order of the 4 li.