Notification with notify.js, project table fix
Downloaded notify.min.js and addd code to notification.js
This is a function that makes a small messages appear (top right of
window by default) to notify the user that his/her action was either
successful or that something went wrong (error).
Called the function in the ui files for all the actions.

The draggable class of the project table was causing bugs so it is back
to being fixed. when adding new projects it would not get bigger and new
project would have to be seen by scrolling down. This would be confusing
for user.