Fix bug: user seeing previous user's resources
A when signing in, a user was able to see a 1 second flash of the
previous user's resources. This was due to the index action not being
immediate which means that previous resources would still be in the
two tables.
Fixed is by creating an empty handlebar file which only has an empty
{{#if}}{{/if}} and injecting that into the 2 tables when signing out.
That way, no matter how fast a user is quick to sign in, nothing will be
displayed but his/her resources if existent.