Implement weather api
Added an input and submit button in the project-table for searching
cities for their current weather.

Added click event to button and added it to the on doc ready.
Added function to fetch data from api (also talks to back end as it
needs to validate key from back-end).

Need to display data and improve visuals.